Updates to provider
