+++ lib/v2/asset_managment/providers/transactions/transaction_provider.dart
@@ -4,53 +4,57 @@ import 'package:ribn/v2/asset_managment/models/asset_type.dart';
 
 // Project imports:
 import 'package:ribn/v2/asset_managment/models/asset_details.dart';
+import 'package:ribn/v2/asset_managment/models/chains.dart';
+import 'package:ribn/v2/asset_managment/providers/selected_network_provider.dart';
 
 final transactionNotifierProvider =
     StateNotifierProvider.autoDispose<TransactionNotifier, AsyncValue<List<AssetDetails>>>((ref) {
-  return TransactionNotifier(ref);
+  final selectedChain = ref.watch(selectedChainProvider);
+  return TransactionNotifier(ref, selectedChain);
 });
 
 class TransactionNotifier extends StateNotifier<AsyncValue<List<AssetDetails>>> {
   final Ref ref;
-  TransactionNotifier(this.ref) : super(AsyncLoading()) {
-    getAssets().then((value) {
-      state = AsyncData(value);
-    });
+  final Chains selectedChain;
+  TransactionNotifier(this.ref, this.selectedChain) : super(AsyncLoading()) {
+    getAssets();
   }
 
-  Future<List<AssetDetails>> getAssets() async {
-    return [
-      AssetDetails(
-        assetName: 'Topl Governance',
-        assetType: AssetType.governance,
-        assetAmount: 8437,
-      ),
-      AssetDetails(
-        assetName: 'Topl Transaction',
-        assetType: AssetType.transaction,
-        assetAmount: 4012,
-      ),
-      AssetDetails(
-        assetName: 'Avalanche',
-        assetType: AssetType.avalanche,
-        assetAmount: 508.6,
-      ),
-      AssetDetails(
-        assetName: 'Polygon',
-        assetType: AssetType.polygon,
-        assetAmount: 0.35,
-      ),
-      AssetDetails(
-        assetName: 'Ethereum',
-        assetType: AssetType.ethereum,
-        assetAmount: 1.53,
-      ),
-      AssetDetails(
-        assetName: 'Bitcoin',
-        assetType: AssetType.bitcoin,
-        assetAmount: 0.35,
-      )
-    ];
+  Future<void> getAssets() async {
+    state = AsyncData(
+      [
+        AssetDetails(
+          assetName: 'Topl Governance',
+          assetType: AssetType.governance,
+          assetAmount: 8437,
+        ),
+        AssetDetails(
+          assetName: 'Topl Transaction',
+          assetType: AssetType.transaction,
+          assetAmount: 4012,
+        ),
+        AssetDetails(
+          assetName: 'Avalanche',
+          assetType: AssetType.avalanche,
+          assetAmount: 508.6,
+        ),
+        AssetDetails(
+          assetName: 'Polygon',
+          assetType: AssetType.polygon,
+          assetAmount: 0.35,
+        ),
+        AssetDetails(
+          assetName: 'Ethereum',
+          assetType: AssetType.ethereum,
+          assetAmount: 1.53,
+        ),
+        AssetDetails(
+          assetName: 'Bitcoin',
+          assetType: AssetType.bitcoin,
+          assetAmount: 0.35,
+        )
+      ],
+    );
   }
 
   Future<List<AssetDetails>> refreshAssets() {
